requirements.txt
requests
urllib3
moviepy
imageio-ffmpeg
Pillow
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
boto3
